refactor dailyHomework source and change css filename
