+++ src/_data/contentDefinition.js
@@ -1,0 +1,15 @@
+const contentKeyKorToEng = {
+    휴식게이지: 'calendar',
+    카오스던전: 'chaosDun',
+    가디언토벌: 'guardianDun',
+    에포나: 'epona',
+    주간가디언: 'guardianDun',
+    어비스던전: 'abyss2types',
+    아르고스: 'argos',
+    발탄: 'baltan',
+    비아키스: 'biakiss',
+    쿠크세이튼: 'kukuseitn',
+    아브렐슈드: 'abrel',
+};
+
+export default contentKeyKorToEng;
